+++ src/Quartz.Serialization.Json/Calendars/BaseCalendarSerializer.cs
@@ -1,0 +1,23 @@
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+
+using Quartz.Impl.Calendar;
+
+namespace Quartz.Calendars
+{
+    internal sealed class BaseCalendarSerializer : CalendarSerializer<BaseCalendar>
+    {
+        protected override BaseCalendar Create(JObject source)
+        {
+            return new BaseCalendar();
+        }
+
+        protected override void SerializeFields(JsonWriter writer, BaseCalendar calendar)
+        {
+        }
+
+        protected override void DeserializeFields(BaseCalendar calendar, JObject source)
+        {
+        }
+    }
+}