+++ changelog.md
@@ -10,6 +10,7 @@
     * Upgrade MySqlConnector to 1.0 (namespace has changed) (#890)
     * Support Microsoft.Extensions.Logging.Abstractions (#756)
     * Support Microsoft.Data.SQLite with full framework (#893)
+    * Support custom calendar JSON serialization (#697)
 
 * FIXES